move: exception 상세 분류
