+++ src/main/java/racingcar/exception/car/DuplicateCarNameException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.car;
 
 public final class DuplicateCarNameException extends IllegalArgumentException {
     public static final String DUPLICATE_CAR_NAME_EXCEPTION_MESSAGE = "중복된 차 이름이 존재합니다";
