+++ src/main/java/racingcar/exception/car/InvalidCarNameLengthException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.car;
 
 public final class InvalidCarNameLengthException extends IllegalArgumentException {
     public static final String INVALID_CAR_NAME_LENGTH_EXCEPTION_MESSAGE = "차 이름은 1~5자여야 합니다.";
