+++ src/main/java/racingcar/exception/car/InvalidCarNamesInputException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.car;
 
 public class InvalidCarNamesInputException extends IllegalArgumentException {
     public static final String INVALID_CAR_NAMES_INPUT = "유효하지 않은 양식입니다.\n차 이름 양식은 ,로 구분해주세요.";
