+++ src/main/java/racingcar/exception/car/NoCarsException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.car;
 
 public final class NoCarsException extends IllegalArgumentException {
     public static final String NO_CARS_EXCEPTION_MESSAGE = "생성할 차가 존재하지 않습니다.";
