+++ src/main/java/racingcar/exception/round/InvalidRoundException.java
@@ -1,4 +1,4 @@
-package racingcar.exception;
+package racingcar.exception.round;
 
 public final class InvalidRoundException extends IllegalArgumentException {
     public static final String INVALID_ROUND_EXCEPTION = "최소 라운드 수는 1입니다";
